backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
qdrant-client>=1.7.0
rank-bm25>=0.2.2
cohere>=4.0.0
requests>=2.31.0
openai>=1.0.0
numpy>=1.24.0
tqdm>=4.65.0
